requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
typing-extensions>=4.8.0
neo4j==5.23.1
structlog>=24.1.0